% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pipeline.R
\name{generate_model}
\alias{generate_model}
\title{Generate a voomNSC model using MLSeq}
\usage{
generate_model(
  data,
  class,
  positive = "T",
  ctrl = MLSeq::voomControl(method = "repeatedcv", number = 5, repeats = 2, tuneLength =
    7)
)
}
\arguments{
\item{data}{A data frame or count matrix (Features as rows, samples as columns)}

\item{class}{A binary vector of strings containing the condition factor}

\item{positive}{The string meaning positive for the class; default "T"}

\item{ctrl}{A control function to modify the model; by default repeatedcv, 2 repeats x 5, tune Length 7}
}
\value{
An MLSeq model that can be passed into use_model
}
\description{
Generate a voomNSC model using MLSeq
}
