% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pipeline.R
\name{generate_model}
\alias{generate_model}
\title{Generate a voomNSC model using MLSeq}
\usage{
generate_model(
  data,
  class,
  positive = "T",
  use_svmRadial = FALSE,
  voom_ctrl = MLSeq::voomControl(method = "repeatedcv", number = 5, repeats = 2,
    tuneLength = 7),
  svm_ctrl = caret::trainControl(method = "repeatedcv", number = 2, repeats = 2,
    classProbs = TRUE),
  train_test_split = FALSE,
  negative = "F",
  normalize = "deseq",
  preProcessing = "deseq-vst"
)
}
\arguments{
\item{data}{A data frame or count matrix (Features as rows, samples as columns)}

\item{class}{A binary vector of strings containing the condition factor}

\item{positive}{The string meaning positive for the class. Note this cannot be "TRUE".}

\item{use_svmRadial}{Determines whether or not an svmRadial model is used instead of voomNSC}

\item{voom_ctrl}{A control function to modify the model for voomNSC models}

\item{svm_ctrl}{A control function to modify the model for SVM models (use only if use_svmRadial is TRUE)}

\item{train_test_split}{Determine whether or not to split (70:30) into train and test data}

\item{negative}{Required if train_test_split is true-- the string meaning negative for the class; default "F"}

\item{normalize}{For voom-based clasifiers. Character string indicating type of normalization-- 'deseq', 'tmm', or 'none'}

\item{preProcessing}{For caret-based classifiers. Name of preprocessing method-- eg "deseq-vst", "deseq-rlog", "logcpm"}
}
\value{
An MLSeq model that can be passed into use_model. If train-test-split is true, then this outputs a named list with components:

\item{model}{The trained model}

\item{confusion_matrix}{The confusion matrix from the tts}
}
\description{
Generate a voomNSC model using MLSeq
}
