% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pipeline.R
\name{use_model}
\alias{use_model}
\title{Use an existing model}
\usage{
use_model(data, model, features = "all", add_one = FALSE)
}
\arguments{
\item{data}{A data frame or count matrix (Features as rows, samples as columns)}

\item{model}{A model generated from generate_model}

\item{features}{The features output from generate_model (ls$features)}

\item{add_one}{Boolean: Adds one to the data to avoid too many zeroes error; helpful if low feature count}
}
\value{
A factor of predictions
}
\description{
Use an existing model
}
