% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pipeline.R
\name{use_model}
\alias{use_model}
\title{Use an existing model}
\usage{
use_model(data, model)
}
\arguments{
\item{data}{A data frame or count matrix (Features as rows, samples as columns)}

\item{model}{A model generated from generate_model}
}
\value{
A factor of predictions
}
\description{
Use an existing model
}
